requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pillow
pytesseract
pdf2image
python-dotenv
google-generativeai
